+++ src/nationalarchives/components/checkboxes/checkboxes.stories.js
@@ -242,9 +242,22 @@ Inline.args = {
 
 export const Single = Template.bind({});
 Single.args = {
+  id: "terms",
+  name: "terms",
+  items: [
+    {
+      text: "I agree to the terms and conditions",
+      value: "agree",
+    },
+  ],
+  classes: "tna-checkboxes--demo",
+};
+
+export const SingleWithLabel = Template.bind({});
+SingleWithLabel.args = {
   label: "Terms and conditions",
   headingLevel: 4,
-  headingSize: "xs",
+  headingSize: "m",
   id: "terms",
   name: "terms",
   items: [